BUG: Change C GLSZM calculation
Change from use of recursive functions to use of while loops. This is to attempt to limit memory usage, as it 'unexpectedly' failed during continuous testing.

Additionally, change datatype for mask to char (signed char in glszm).
Nd in gldm is equal to Na, remove passing of Nd as argument.