BUG: Fix bug in GLSZM calculation in C
Fix potential index out of range error in `fill_glszm` in tempData. This occurs when the entire tempData matrix is filled (Ns different regions. When the function fill_glszm then runs, it does not exit the while loop (as it encounters no -1 at the end of the matrix).
Specify the size of the tempData matrix = (Ns + 1) * 2, so that the last element will always be -1 (max idx computed by function calculate_glszm is still Ns * 2).

Additionally re-apply code style to C files